refactor(internal): refactor CSV traversal logic
