+++ test.ts
@@ -3,8 +3,9 @@ import {withCSV} from './src'
 async function main() {
   const test = await withCSV('./small.csv')
     .columns(['City'])
-    .filter(() => {
-      return true
+    .map(row => {
+      throw new Error('SIKE')
+      return row
     })
     .rows()
